fix(write) Fix to ensure existing TTL remains for kv
If the previous value was supplied, but a new TTL wasn't then the new expiry was based on the defaultLockTTL, which is incorrect. It now uses the existing value.
